further separated stats from live by moving over _base.py and having live endpoints reference that
