+++ nba_api/live/endpoints/playbyplay.py
@@ -1,4 +1,4 @@
-from nba_api.stats.endpoints._base import Endpoint
+from nba_api.live.endpoints._base import Endpoint
 from nba_api.live.library.http import NBAStatsHTTP
 class PlayByPlay(Endpoint):
     endpoint_url = 'playbyplay/playbyplay_{game_id}.json'
